Use of async email send, logic to record network data and create right folders and read back the data and send emails
Email client would close before the async task finished so changed it to follow async/await top to bottom. Traceroute change to use a list for future changes to move it to async patterns as well to avoid blocking.

Added logic to determine right data folder structure, save data to file there and read back that data at end of day/week in order to send an email. Implemented sending of daily emails